Combine client binaries, use flag for codec
