+++ interop/src/bin/client.rs
@@ -1,4 +1,5 @@
-use interop::client;
+use interop::client::{InteropTest, InteropTestUnimplemented};
+use interop::{client_prost, client_protobuf};
 use std::{str::FromStr, time::Duration};
 use tonic::transport::Endpoint;
 use tonic::transport::{Certificate, ClientTlsConfig};
@@ -7,6 +8,25 @@ use tonic::transport::{Certificate, ClientTlsConfig};
 struct Opts {
     use_tls: bool,
     test_case: Vec<Testcase>,
+    codec: Codec,
+}
+
+#[derive(Debug)]
+enum Codec {
+    Prost,
+    Protobuf,
+}
+
+impl FromStr for Codec {
+    type Err = String;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s {
+            "prost" => Ok(Codec::Prost),
+            "protobuf" => Ok(Codec::Protobuf),
+            _ => Err(format!("Invalid codec: {}", s)),
+        }
+    }
 }
 
 impl Opts {
@@ -17,6 +37,7 @@ impl Opts {
             test_case: pargs.value_from_fn("--test_case", |test_case| {
                 test_case.split(',').map(Testcase::from_str).collect()
             })?,
+            codec: pargs.value_from_str("--codec")?,
         })
     }
 }
@@ -48,8 +69,19 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     let channel = endpoint.connect().await?;
 
-    let mut client = client::TestClient::new(channel.clone());
-    let mut unimplemented_client = client::UnimplementedClient::new(channel);
+    let (mut client, mut unimplemented_client): (
+        Box<dyn InteropTest>,
+        Box<dyn InteropTestUnimplemented>,
+    ) = match matches.codec {
+        Codec::Prost => (
+            Box::new(client_prost::TestClient::new(channel.clone())),
+            Box::new(client_prost::UnimplementedClient::new(channel)),
+        ),
+        Codec::Protobuf => (
+            Box::new(client_protobuf::TestClient::new(channel.clone())),
+            Box::new(client_protobuf::UnimplementedClient::new(channel)),
+        ),
+    };
 
     let mut failures = Vec::new();
 
@@ -58,31 +90,25 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
         let mut test_results = Vec::new();
 
         match test_case {
-            Testcase::EmptyUnary => client::empty_unary(&mut client, &mut test_results).await,
-            Testcase::LargeUnary => client::large_unary(&mut client, &mut test_results).await,
-            Testcase::ClientStreaming => {
-                client::client_streaming(&mut client, &mut test_results).await
-            }
-            Testcase::ServerStreaming => {
-                client::server_streaming(&mut client, &mut test_results).await
-            }
-            Testcase::PingPong => client::ping_pong(&mut client, &mut test_results).await,
-            Testcase::EmptyStream => client::empty_stream(&mut client, &mut test_results).await,
+            Testcase::EmptyUnary => client.empty_unary(&mut test_results).await,
+            Testcase::LargeUnary => client.large_unary(&mut test_results).await,
+            Testcase::ClientStreaming => client.client_streaming(&mut test_results).await,
+            Testcase::ServerStreaming => client.server_streaming(&mut test_results).await,
+            Testcase::PingPong => client.ping_pong(&mut test_results).await,
+            Testcase::EmptyStream => client.empty_stream(&mut test_results).await,
             Testcase::StatusCodeAndMessage => {
-                client::status_code_and_message(&mut client, &mut test_results).await
+                client.status_code_and_message(&mut test_results).await
             }
             Testcase::SpecialStatusMessage => {
-                client::special_status_message(&mut client, &mut test_results).await
+                client.special_status_message(&mut test_results).await
             }
-            Testcase::UnimplementedMethod => {
-                client::unimplemented_method(&mut client, &mut test_results).await
-            }
+            Testcase::UnimplementedMethod => client.unimplemented_method(&mut test_results).await,
             Testcase::UnimplementedService => {
-                client::unimplemented_service(&mut unimplemented_client, &mut test_results).await
+                unimplemented_client
+                    .unimplemented_service(&mut test_results)
+                    .await
             }
-            Testcase::CustomMetadata => {
-                client::custom_metadata(&mut client, &mut test_results).await
-            }
+            Testcase::CustomMetadata => client.custom_metadata(&mut test_results).await,
             _ => unimplemented!(),
         }
